Implement C_CombatWeaponClone
C_CombatWeaponClone is a class in Portal 2 that was used to fix worldmodels not rendering through portals. This wasn't in any leak but was instead reverse engineered by me, there are probably some "inaccuracies" though.
One strange bug is that going into thirdperson and going back to firstperson will make the model clone not render. The current implementation is good enough for now though.